% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flags.R
\name{add_flags}
\alias{add_flags}
\title{Calculate flags for a set of records and add them to the provided data frame}
\usage{
add_flags(indf, show_summary = TRUE, quiet = FALSE, ...)
}
\arguments{
\item{indf}{Required. Properly formatted data frame containing a row per
record}

\item{show_summary}{Optional. Show a summary of the quality flags after the
process has finished. Defaults to TRUE}

\item{quiet}{Optional. Don't show any logging message at all. Defaults to
FALSE}

\item{...}{Any extra parameters for \code{httr} \code{\link{POST}}}
}
\value{
The provided data frame with the quality flags added as new columns
}
\description{
\code{add_flags} calls the \code{POST} method of the API in order to extract
the flags for a set of records. After finishing, the function returns the
provided data.frame with a new column, \code{flags}, which holds for each
record a list of the geospatial quality flags. It also shows a summary of the
results, indicating how many records showed different types of issues.
\strong{NOTE:} currently, the API imposes a hard-limit of 1000 records per
request, to avoid malfunctioning due to some third-party library limtations.
}
\examples{
\dontrun{
# Using the rgbif package
if (requireNamespace("rgbif", quietly=TRUE)) {
 library("rgbif")

 # Prepare data
 d <- occ_data(scientificName="Apis mellifera", limit=50, minimal=FALSE)
 d <- d$data

 # Format data.frame
 d <- format_gq(d, source="rgbif")

 # Execute the call to the API, showing output and
 # logging information, and store the results
 dd <- add_flags(d)

 # Execute the call without showing summary output, but
 # showing logging information
 dd <- add_flags(d, show_summary=FALSE)

 # Execute the call without showing any logging at all
 # (except errors, obviously)
 dd <- add_flags(d, quiet=TRUE)

 # Data quality output will be stored in a new field called flags
 names(dd$flags)

 # You can check records with certain flags as usual
 # See records with coordinates-country mismatch
 dd[dd$flags$coordinatesInsideCountry == FALSE,]
}
}

}
\seealso{
\code{\link{format_gq}}, \code{\link{flags}},
  \code{\link{parse_record}}
}

