% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{flags}
\alias{flags}
\title{Calculate flags for a single record}
\usage{
flags(record = NA, decimalLatitude = NA, decimalLongitude = NA,
  countryCode = "", scientificName = "")
}
\arguments{
\item{record}{List-type object containing information on the record. If
present, it MUST contain the following four attributes as named elements.}

\item{decimalLatitude}{Only if 'record' is not present. Latitude in decimal
degrees format (float, e.g. 42.1833)}

\item{decimalLongitude}{Only if 'record' is not present. Longitude in decimal
degrees format (float, e.g. -1.8332)}

\item{countryCode}{Only if 'record' is not present. Two or three-letter
ISO-3166 country codes (string, e.g. "ES")}

\item{scientificName}{Only if 'record' is not present. Binomial identifying
the species (string, e.g. "Puma concolor")}
}
\value{
A named list with the geospatial quality flags
}
\description{
\code{flags} calls the \code{GET} method of the API in order to extract
the flags for an individual record. It returns just the \code{flags} element
}
\examples{
rec <- list(decimalLatitude=42.1833, decimalLongitude=-1.8332,
            countryCode="ES", scientificName="Puma concolor")
flags(record=rec)

# OR

flags(decimalLatitude=42.1833, decimalLongitude=-1.8332,
           countryCode="ES", scientificName="Puma concolor")

}
\seealso{
\code{\link{add_flags}}
}

