% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multi.R
\name{get_multi}
\alias{get_multi}
\title{Calculate flags for a set of records}
\usage{
get_multi(indf)
}
\arguments{
\item{indf}{Required. Properly formatted data frame containing a row per
record (see below)}
}
\value{
The provided data frame with the quality flags added as new columns
}
\description{
\code{get_multi} calls the \code{POST} method of the API in order to extract
the flags for a set of records
}
\seealso{
\code{\link{gq_format}}, \code{\link{get_single}}
}

