% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_single.R
\name{get_single}
\alias{get_single}
\title{Calculate flags for a single record}
\usage{
get_single(record = NA, decimalLatitude = NA, decimalLongitude = NA,
  countryCode = "", scientificName = "")
}
\arguments{
\item{record}{Vector-type object containing information on the record. If
present, it MUST contain the following four named elements.}

\item{decimalLatitude}{Only if 'record' is not present. Latitude in decimal
degrees format (float, e.g. 42.1833)}

\item{decimalLongitude}{Only if 'record' is not present. Longitude in decimal
degrees format (float, e.g. -1.8332)}

\item{countryCode}{Only if 'record' is not present. Two or three-letter
ISO-3166 country codes (string, e.g. "ES")}

\item{scientificName}{Only if 'record' is not present. Binomial identifying
the species (string, e.g. "Puma concolor")}
}
\value{
A vector with the supplied record plus the geospatial quality flags
  in the $flags element
}
\description{
\code{get_single} calls the \code{GET} method of the API in order to extract
the flags for an individual record
}
\examples{
rec <- list(decimalLatitude=42.1833, decimalLongitude=-1.8332,
            countryCode="ES", scientificName="Puma concolor")
get_single(record=rec)

# OR

get_single(decimalLatitude=42.1833, decimalLongitude=-1.8332,
           countryCode="ES", scientificName="Puma concolor")
}

