% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gq_format.R
\name{gq_format}
\alias{gq_format}
\title{Rename fields and modify content type to fit the geospatial quality API
format}
\usage{
gq_format(indf)
}
\arguments{
\item{indf}{Required. Data frame containing a row per record}
}
\value{
The provided data frame with names changed to fit the geospatial
  quality API format
}
\description{
\code{gq_format} performs two actions on the provided data frame. First, it
attempts to change the names of certain fields (coordinates, country code and
scientific name) for the ones accepted by the API. Second, it parses the
content of these fields and modifies the format of the values if needed (e.g.
turns coordinates into floating point numbers).
}
\examples{
records <- gq_format(records)

}
\seealso{
\code{\link{get_multi}}
}

