% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flags.R
\name{add_flags}
\alias{add_flags}
\title{Calculate flags for a set of records and add them to the provided data frame}
\usage{
add_flags(indf = NA, show_summary = TRUE, silent = FALSE)
}
\arguments{
\item{indf}{Required. Properly formatted data frame containing a row per
record}

\item{show_summary}{Optional. Show a summary of the quality flags after the
process has finished. Defaults to TRUE}

\item{silent}{Optional. Don't show any logging message at all. Defaults to
FALSE}
}
\value{
The provided data frame with the quality flags added as new columns
}
\description{
\code{add_flags} calls the \code{POST} method of the API in order to extract
the flags for a set of records
}
\examples{
# Using the rgbif package, prepare data
d <- rgbif::occ_data(scientificName="Apis mellifera", limit=50, minimal=FALSE)
d <- d$data

# Execute the call to the API and store the results
dd <- add_flags(d)

}
\seealso{
\code{\link{flags}}
}

