% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_single.R
\name{get_single}
\alias{get_single}
\title{Calculate flags for a single record}
\usage{
get_single(decimalLatitude = NA, decimalLongitude = NA, countryCode = NA,
  scientificName = "")
}
\arguments{
\item{decimalLatitude}{Latitude in decimal degrees format (float, e.g.
42.1833)}

\item{decimalLongitude}{Longitude in decimal degrees format (float, e.g.
-1.8332)}

\item{countryCode}{Two or three-letter ISO-3166 country codes (string, e.g.
"ES")}

\item{scientificName}{Binomial identifying the species (string, e.g. "Puma
concolor")}
}
\value{
The supplied record wih the geospatial quality flags
}
\description{
\code{get_single} calls the \code{GET} method of the API in order to extract
the flags for an individual record
}
\examples{
get_single(decimalLatitude=42.1833, decimalLongitude=-1.8332, countryCode="ES", scientificName="Puma concolor")
}
\seealso{
\code{\link{http://bioinformatics.oxfordjournals.org/content/early/2016/02/16/bioinformatics.btw057}}
  for more info on the API
}

